cmus lib made fully async.
When several cmus requests are made in quick succession (e.g., when the
UI tries to dequeue a half dozen songs at once), you get interleaving of
requests and responses. So if the response to the first dequeue request
actually manages to make it back to the client last, it will appear as
though most of your dequeues had no effect. Furthermore, the way the
dequeue requests are implemented (by clearing the queue then re-adding
the stuff we want to leave) means that if the re-adds are done out of
order, it won't just look ineffective; it will actually be ineffective.

This called for a system that was async on the frontend, but forced all
calls to cmus to be serialized. The way it works is:
Request#1: create promise#1. LastCmd = promise#1. Return LastCmd.
Request#2: create promise#2. LastCmd = LastCmd.then(promise#2). Return
LastCmd.
Request#3: create promise#3. LastCmd = LastCmd.then(promise#3). Return
LastCmd.

And so on. This means none of the cmus lib functions actually do
anything immediately. They always return functions that make a promise,
and that function is called when the _last_ function's promise
completes. So the scenario above is slightly simplified: instead of
creating promise#2, Request#2 actually creates a function that creates
promise#2.