static, named queues.
Rather than the queue being consumed as it plays, the backend provides
the illusion of a static queue, with one named queue per 'user'.

TODO:
- Show 'now playing' in queue
- Ability to create new named queues? Or does this encroach on
playlists?
- Should all playlists just be converted to named queues? Would allow us
to retain metadata like year, duration, etc. Yeah, this is _probably_
best. A tool to generate m3us (or a server command to do so) would
maintain compatibility with my phone-transfer scripts. Or, could just
let the server drive all of that directly. hrm.