cmus 'main' playlist; play button everywhere
We're now using the cmus playlist to track the main songlist of the ui.
This has the advantage of making it possible to use the built-in
playlist/queue features of cmus to provide the same abstraction via the
ui. Prior to this commit, we just used the cmus queue to model
everything we did, but that has some pretty fundamental issues wrt the
ui knowing when a consumed track belonged to the 'upnext' queue or the
playlist.

It's definitely duct tape and twine to make the playlist refresh work.
Cmus now supports multiple playlists, which is great but it seems to
break the `cmus-remote -c path/to/playlist.m3u` command line, so we
have to work around it, and I think it works as long as you don't open
cmus and use tab to change the focus from the playlist list to the
playlist songs.