Data model reorg
I started writing out how I wanted the data model to work in
nodes/readme.txt, then I started doing it.

The overall idea is that since the ui can only show things that the
server gives it, we should be able to remove all path parsing at the
server level. The tricky part is how to handle the fact that for us to
efficiently convert paths to a 'proper' song (i.e., with metadata beyond
what you can parse out of the path), we need to keep the song in memory
(it takes beet ~1000ms to lookup a track by path). But obviously, we
don't want to just store the entire database in memory, particularly on
our tiny little pi. So instead, we cache stuff as we give it back to the
UI, and we maintain a small 'history' of caches, on the theory that the
ui will only ever make requests with paths it has recently asked for.

This required converting all path-based metadata (basically, the artist
cache and the m3u playlists) to the two formats we trust: a 'parseable
line' format (this is the format in which queries fall out of beet) or
json-stringified versions of our in-memory songs.

This change also stops the practice of the ui having the entire tag
database in memory. This means it has to make a request to the server
when it wants to load songs, but it has thus far been very quick since
the server keeps the tags database in memory.

The 'static' or 'named' queues are currently broken. to be fixed soon.
cmus now treats them as just another playlist; it is the UI that will
provide the illusion of a 'current' queue. We will refer to those as
'stashes' from now on, and it will be a strictly ui concept.